refactor(mvm-build): migrate JobResult + read_job_result into BuilderVmRuntime
Plan 97 Phase C, second migration after stage_job_dir (#434). Lifts the
hypervisor-agnostic /job/result JSON parsing out of libkrun_builder.rs
into builder_vm_runtime.rs so the future VzBuilderVm path can reuse it
verbatim. JobResult and read_job_result are pub (matching stage_job_dir)
since cross-crate VzBuilderVm consumers will need them and dead-code
analysis ignores pub items in default-feature builds.

Test coverage: the three existing tests move alongside the function;
adds a fourth test pinning the ExtractionFailed arm on malformed JSON,
which the libkrun-side suite never exercised.
